# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: leads
    description: Lead operations
  - name: conversations
    description: Conversation operations
  - name: integrations
    description: Integration operations
  - name: webhooks
    description: Platform webhook receiver and settings
  - name: billing
    description: Stripe billing operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      description: Returns the leads visible to the current workspace
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead
      description: Returns a single lead by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Lead not found
  /leads/{id}/advance:
    post:
      operationId: advanceLead
      tags: [leads]
      summary: Advance a demo lead stage
      description: Advances a demo lead to the next pipeline stage. Real stage changes come from verified platform or website events.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "409":
          description: Lead is not in demo mode or is already at the final stage
  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List conversations
      description: Returns the conversations visible to the current workspace
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
  /conversations/{id}:
    get:
      operationId: getConversation
      tags: [conversations]
      summary: Get a conversation
      description: Returns a single conversation by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
          description: Conversation not found
  /conversations/{id}/approve:
    post:
      operationId: approveConversation
      tags: [conversations]
      summary: Approve an AI-suggested reply
      description: Approves a reply for the conversation. The reply is stored and flagged approved; it is only ever delivered after the platform adapter confirms a real send through an official API. Never simulates a send.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reply]
              properties:
                reply:
                  type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApproveConversationResponse"
        "400":
          description: Reply is empty or invalid
        "404":
          description: Conversation not found
  /integrations:
    get:
      operationId: listIntegrations
      tags: [integrations]
      summary: List integrations
      description: Returns available social platform integrations and their connection state
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Integration"
  /integrations/{id}/connect:
    post:
      operationId: connectIntegration
      tags: [integrations]
      summary: Get an OAuth authorization URL
      description: Returns the official OAuth authorization URL for a platform. Credentials must be configured via environment variables.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionResult"
        "404":
          description: Integration not found
        "409":
          description: Integration is already connected
        "503":
          description: OAuth credentials are not configured
  /integrations/oauth/callback:
    get:
      operationId: oauthCallback
      tags: [integrations]
      summary: OAuth callback
      description: Handles the OAuth redirect after a user authorizes a platform
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Connected
        "400":
          description: Authorization failed or missing parameters
        "501":
          description: Token exchange not implemented for this platform
  /settings/webhooks:
    get:
      operationId: listWebhookSettings
      tags: [webhooks]
      summary: List webhook settings
      description: Returns per-platform webhook callback URLs and verify tokens for registering in platform developer dashboards
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSettingsResponse"
  /settings/webhooks/events:
    get:
      operationId: listWebhookEvents
      tags: [webhooks]
      summary: List recent webhook events
      description: Returns the most recent webhook events received from platforms
      parameters:
        - name: platform
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookEvent"
  /webhooks/{platform}:
    get:
      operationId: verifyWebhook
      tags: [webhooks]
      summary: Platform webhook verification
      description: Answers a platform webhook verification challenge
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Challenge echoed
        "403":
          description: Verification failed
    post:
      operationId: receiveWebhook
      tags: [webhooks]
      summary: Receive a platform webhook event
      description: Stores an incoming platform webhook event after token verification
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
        - name: token
          in: query
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/WebhookTokenHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
        "401":
          description: Invalid webhook token
  /billing/subscription:
    get:
      operationId: getSubscription
      tags: [billing]
      summary: Current subscription
      description: Returns the workspace subscription plan and status
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"
  /billing/checkout:
    post:
      operationId: createCheckout
      tags: [billing]
      summary: Create a Stripe Checkout session
      description: Creates a Stripe subscription Checkout Session and returns its URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan]
              properties:
                plan:
                  $ref: "#/components/schemas/Plan"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    nullable: true
        "400":
          description: Invalid plan
        "503":
          description: Stripe or database not configured
  /billing/webhook:
    post:
      operationId: billingWebhook
      tags: [billing]
      summary: Stripe webhook
      description: Receives Stripe subscription events and syncs the local subscription row
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
      parameters:
        - $ref: "#/components/parameters/StripeSignatureHeader"
components:
  parameters:
    WebhookTokenHeader:
      name: x-webhook-token
      in: header
      required: false
      schema:
        type: string
    StripeSignatureHeader:
      name: stripe-signature
      in: header
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DataMode:
      type: string
      enum:
        - DEMO
        - REAL
        - NOT_CONNECTED
    LeadStatus:
      type: string
      enum:
        - DISCOVERED
        - ENGAGED
        - INTERESTED
        - QUALIFIED
        - WEBSITE VISIT
        - CONVERSION
        - CUSTOMER
    Plan:
      type: string
      enum:
        - starter
        - pro
    Lead:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        username:
          type: string
        platform:
          type: string
        source:
          type: string
        campaign:
          type: string
        interaction:
          type: string
        history:
          type: string
        leadScore:
          type: integer
        intentScore:
          type: integer
        status:
          $ref: "#/components/schemas/LeadStatus"
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
        lastInteraction:
          type: string
        nextFollowUp:
          type: string
        conversionStatus:
          type: string
        mode:
          $ref: "#/components/schemas/DataMode"
      required:
        - id
        - name
        - username
        - platform
        - source
        - campaign
        - interaction
        - history
        - leadScore
        - intentScore
        - status
        - tags
        - notes
        - lastInteraction
        - nextFollowUp
        - conversionStatus
        - mode
    Conversation:
      type: object
      properties:
        id:
          type: string
        leadId:
          type: string
        name:
          type: string
        username:
          type: string
        platform:
          type: string
        preview:
          type: string
        intent:
          type: string
        sentiment:
          type: string
        status:
          type: string
        suggestion:
          type: string
        replyText:
          type: string
        replyStatus:
          type: string
          enum:
            - pending
            - approved
            - sent
        sentAt:
          type: string
        mode:
          $ref: "#/components/schemas/DataMode"
      required:
        - id
        - leadId
        - name
        - username
        - platform
        - preview
        - intent
        - sentiment
        - status
        - suggestion
        - mode
    ApproveConversationResponse:
      type: object
      properties:
        conversation:
          $ref: "#/components/schemas/Conversation"
        delivery:
          type: string
          enum:
            - requires_connection
            - queued
      required:
        - conversation
        - delivery
    Integration:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        connected:
          type: boolean
      required:
        - id
        - name
        - description
        - icon
        - connected
    ConnectionResult:
      type: object
      properties:
        id:
          type: string
        authUrl:
          type: string
        state:
          type: string
      required:
        - id
        - authUrl
        - state
    WebhookSetting:
      type: object
      properties:
        platform:
          type: string
        callbackUrl:
          type: string
        verifyToken:
          type: string
        requiresWebhook:
          type: boolean
        configured:
          type: boolean
      required:
        - platform
        - callbackUrl
        - verifyToken
        - requiresWebhook
        - configured
    WebhookSettingsResponse:
      type: object
      properties:
        base:
          type: string
        settings:
          type: array
          items:
            $ref: "#/components/schemas/WebhookSetting"
      required:
        - base
        - settings
    WebhookEvent:
      type: object
      properties:
        id:
          type: integer
        platform:
          type: string
        eventType:
          type: string
        receivedAt:
          type: string
      required:
        - id
        - platform
        - eventType
        - receivedAt
    SubscriptionStatus:
      type: object
      properties:
        plan:
          type: string
        status:
          type: string
        currentPeriodEnd:
          type: string
          nullable: true
        stripeConfigured:
          type: boolean
      required:
        - plan
        - status
        - stripeConfigured